updated examples to use uint32_t and uint64_t

+++ examples/bluescope/testbluescope.cpp
@@ -74,10 +74,10 @@ public:
   MemcpyIndication(unsigned int id) : MemcpyIndicationWrapper(id){}
 
 
-  virtual void started(unsigned long words){
+  virtual void started(uint32_t words){
     fprintf(stderr, "started: words=%ld\n", words);
   }
-  virtual void done(unsigned long mismatch) {
+  virtual void done(uint32_t mismatch) {
     sem_post(&done_sem);
     finished = true;
     memcmp_fail |= mismatch;
@@ -91,21 +91,21 @@ public:
       //dump("dbg", (char*)bsBuffer,  128);   
     // }
   }
-  virtual void rData ( unsigned long long v ){
+  virtual void rData ( uint64_t v ){
     //fprintf(stderr, "rData: %016llx\n", v);
   }
-  virtual void readReq(unsigned long v){
+  virtual void readReq(uint32_t v){
     //fprintf(stderr, "readReq %lx\n", v);
   }
-  virtual void writeReq(unsigned long v){
+  virtual void writeReq(uint32_t v){
     //fprintf(stderr, "writeReq %lx\n", v);
   }
-  virtual void writeAck(unsigned long v){
+  virtual void writeAck(uint32_t v){
     //fprintf(stderr, "writeAck %lx\n", v);
   }
-  virtual void reportStateDbg(unsigned long streamRdCnt, 
-			      unsigned long streamWrCnt, 
-			      unsigned long dataMismatch){
+  virtual void reportStateDbg(uint32_t streamRdCnt, 
+			      uint32_t streamWrCnt, 
+			      uint32_t dataMismatch){
     fprintf(stderr, "Memcpy::reportStateDbg: streamRdCnt=%ld, streamWrCnt=%ld, dataMismatch=%ld\n", 
 	    streamRdCnt, streamWrCnt, dataMismatch);
   }  
@@ -121,7 +121,7 @@ public:
     fprintf(stderr, "BlueScope::triggerFired\n");
     trigger_fired = true;
   }
-  virtual void reportStateDbg(unsigned long long mask, unsigned long long value){
+  virtual void reportStateDbg(uint64_t mask, uint64_t value){
     //fprintf(stderr, "BlueScope::reportStateDbg mask=%016llx, value=%016llx\n", mask, value);
     fprintf(stderr, "BlueScope::reportStateDbg\n");
     dump("    mask =", (char*)&mask, sizeof(mask));
