+++ examples/memread128/testmemread.cpp
@@ -32,25 +32,25 @@ class MemreadIndication : public MemreadIndicationWrapper
 {
 public:
   unsigned int rDataCnt;
-  virtual void readReq(unsigned long v){
+  virtual void readReq(uint32_t v){
     fprintf(stderr, "Memread::readReq %lx\n", v);
   }
-  virtual void readDone(unsigned long v){
+  virtual void readDone(uint32_t v){
     fprintf(stderr, "Memread::readDone mismatch=%lx\n", v);
     mismatchCount += v;
     sem_post(&test_sem);
   }
-  virtual void started(unsigned long words){
+  virtual void started(uint32_t words){
     fprintf(stderr, "Memread::started: words=%lx\n", words);
   }
-  virtual void rData ( unsigned long long v ){
+  virtual void rData ( uint64_t v ){
     fprintf(stderr, "rData (%08x): ", rDataCnt++);
     dump("", (char*)&v, sizeof(v));
   }
-  virtual void reportStateDbg(unsigned long streamRdCnt, unsigned long dataMismatch){
+  virtual void reportStateDbg(uint32_t streamRdCnt, uint32_t dataMismatch){
     fprintf(stderr, "Memread::reportStateDbg: streamRdCnt=%08lx dataMismatch=%ld\n", streamRdCnt, dataMismatch);
   }  
-  virtual void mismatch(unsigned long offset, unsigned long long ev, unsigned long long v) {
+  virtual void mismatch(uint32_t offset, uint64_t ev, uint64_t v) {
     fprintf(stderr, "Mismatch at %lx %llx != %llx\n", offset, ev, v);
     mismatchesReceived++;
   }
@@ -107,8 +107,8 @@ int main(int argc, const char **argv)
 #endif
   device->startRead(ref_srcAlloc, numWords, burstLen, iterCnt);
   sem_wait(&test_sem);
-  unsigned long long cycles = lap_timer(0);
-  unsigned long long beats = dma->show_mem_stats(ChannelType_Read);
+  uint64_t cycles = lap_timer(0);
+  uint64_t beats = dma->show_mem_stats(ChannelType_Read);
 
   fprintf(stderr, "memory read utilization (beats/cycle): %f\n", ((float)beats)/((float)cycles));
