+++ examples/ring/testring.cpp
@@ -152,12 +152,12 @@ void dump(const char *prefix, char *buf, size_t len)
 class RingIndication : public RingIndicationWrapper
 {
 public:
-  virtual void setResult(long unsigned int cmd, long unsigned int regist, long long unsigned int addr) {
+  virtual void setResult(uint32_t cmd, uint32_t regist, uint64_t addr) {
     fprintf(stderr, "setResult(cmd %ld regist %ld addr %llx)\n", 
 	    cmd, regist, addr);
     sem_post(&setresult_sem);
   }
-  virtual void getResult(long unsigned int cmd, long unsigned int regist, long long unsigned int addr) {
+  virtual void getResult(uint32_t cmd, uint32_t regist, uint64_t addr) {
     fprintf(stderr, "getResult(cmd %ld regist %ld addr %llx)\n", 
 	    cmd, regist, addr);
     /* returning query about last pointer of cmd ring */
